+++ service/src/main/java/com/perfulandia/service/Inventory/exception/ProductoNotFoundException.java
@@ -1,4 +1,4 @@
-package com.perfulandia.service.inventory.exception;
+package com.perfulandia.service.Inventory.exception;
 
 public class ProductoNotFoundException extends RuntimeException {
     public ProductoNotFoundException(String mensaje) {
