+++ service/src/main/java/com/perfulandia/service/Inventory/model/Producto.java
@@ -1,15 +1,25 @@
-package com.perfulandia.service.inventory.dto;
+package com.perfulandia.service.Inventory.model;
 
+import jakarta.persistence.*;
 import lombok.*;
 
+@Entity
+@Table(name = "productos")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ProductoDTO {
+public class Producto {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     private String nombre;
+
     private String descripcion;
+
     private Integer stock;
+
     private Double precio;
 }
