+++ service/src/main/java/com/perfulandia/service/Inventory/repository/ProductoRepository.java
@@ -1,6 +1,6 @@
-package com.perfulandia.service.inventory.repository;
+package com.perfulandia.service.Inventory.repository;
 
-import com.perfulandia.service.inventory.model.Producto;
+import com.perfulandia.service.Inventory.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductoRepository extends JpaRepository<Producto, Long> {
